Extract dataclasses to own module
